cc_blink: Remove WebImageLayer and WebLayerImplFixedBounds
Move the functionality of WebLayerImplFixedBounds directly to
GraphicsLayer. When SetContentsToImage() is called, it stores
the size of the image. Then when UpdateContentRect() occurs,
the image_layer_ (which is the contents layer) is resized
always to match the image, and transformed to fill the
content_rect.

The PaintArtifactCompositor was mutating the contents layer
bounds which would break this path, but it didn't intend
to do so for layers it doesn't control. So we move bounds
setting out of PaintArtifactCompositor for foreign layers,
and set the bounds and DrawsContent(true) at the sites where
the ForeignLayerDisplayItem is created when appropriate (ie
for each case except for GraphicsLayer::ContentsLayer()).

With the removal of WebLayerImplFixedBounds there are no
tests left in cc_blink_unittests, so we remove the test
suite.

R=trchen@chromium.org

Bug: 838693
Cq-Include-Trybots: luci.chromium.try:android_optional_gpu_tests_rel;master.tryserver.blink:linux_trusty_blink_rel;master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2
Change-Id: I2fee4f382eb450af4646a087a08806798a975dd7
Reviewed-on: https://chromium-review.googlesource.com/1053765
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Tien-Ren Chen <trchen@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#557946}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c654ee39b1ec41d9199a49919c8206f9261bb027