Add Stability.ChildFrameCrash.Visibility for tracking sad *sub*frames.
This CL adds Stability.ChildFrameCrash.Visibility which is logged after
detecting that a sad subframe is shown.  The logged enum describes the
timing of a renderer crash in relation to whether the subframe is
visible or not.

Bug: 841493
Change-Id: Ic893764817bd1a9aa31684d53f858a398a31dfc8
Reviewed-on: https://chromium-review.googlesource.com/1054099
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#559356}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 873e60893e31e6d54996ea07d954990133529f79