Add blocking insecure downloads behind a flag
This adds a feature flag that when enabled blocks downloads of unsafe
types when either the final download origin or any origin in the
redirect chain leading to the download are insecure, and the download
file type is potentially unsafe (that is, of mime-type "application/*").
A console warning is logged when a download is blocked in this way.

This behavior is disabled by default.

Bug: 839986
Change-Id: I9a91c82ea043e4bdec5b33e01dc6611734b57a39
Reviewed-on: https://chromium-review.googlesource.com/1045253
Commit-Queue: Christopher Thompson <cthomp@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#560089}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2ffffa2c9a8e9c70ceeed46b57f243aadee85970