Collect metric for the number of clients we asked in async targeting.
Record the number of clients we asked when doing async targeting. 0
means we didn't take async path at all; 1 means we found the target
when asking the first client; 2 means we found the target when asking
the second client; ... This should give us a better idea of how many
clients we usually ask in async path so we can try to optimize and
reduce latency in async path.

Note that we only report this UMA if Site Isolation is turned on and
we actually did explicit hit-testing for that event.

Bug: 732417
Change-Id: I164a9f8c853181c4ab6896a5bfcf9c0806a3da7d
Reviewed-on: https://chromium-review.googlesource.com/1060286
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Ria Jiang <riajiang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#560198}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 69763ca95f4b19782bca7614c073b25d5671aa53