Use more robust emulation_server start process
Uses two threads, one to start the server, and another to 'ping' it
until it connects to a socket. Note that connecting on a socket is a
safe operation that should always succeed, even if the handler is a
blackhole.

Bug: 823776
Change-Id: Ibee24eca923a5f0b860241b1204791392ac408b1
Reviewed-on: https://chromium-review.googlesource.com/1069417
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#560740}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 30c13db2e2abaed7ad430c2582ad4c784086c5e7