Enable suspicious site trigger to run in monitor mode.
In this mode the trigger just listens for certain events and updates
metrics, but never sends reports. This is useful to track the frequency
of suspicious site hits even if reporting quota is spent for the user.

Bug: 817377
Change-Id: Ie524953a438e41a42a63d1e1852f1cfc3093f353
Reviewed-on: https://chromium-review.googlesource.com/1065575
Commit-Queue: Luke Z <lpz@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Jialiu Lin <jialiul@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#560766}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 519d9521d16d9d3af3036daf4d1d5f4398f4396a