Hide the ambient install infobar for 90 days if it has been dismissed.
This CL reduces the annoyingness of the ambient badge infobar on Android
when browsing a PWA. If the infobar is dismissed, it will not be shown
for 90 days, mirroring the existing block time for infobar app banners.
The infobar will continue to be shown on navigations until it is
dismissed.

BUG=782120,834259

Change-Id: Idbbb9e9451fd01f639e859a1b65470330929f993
Reviewed-on: https://chromium-review.googlesource.com/1068903
Reviewed-by: Ben Wells <benwells@chromium.org>
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#560977}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b423ee3e193af2ec44260d1f19590fa5b7ea9980