[Supersize] Fix Disassemble() and store linker_name in metadata.
Disassemble() previously fails because in Supersize Console,
ToolPrefixFinder has linker_name = None. However, in Supersize Archive,
linker_name is parsed from the .map file. So to make it available for
Supersize Console, we should store it in as metadata in the .size file.
It was then discovered that llvm-objdump does not support the
'--demangle' option. So, ToolPrefixFinder.IsLld() is added, and we
exclude '--demangle' for LLD.

Bug: 848731
Change-Id: Ie16110d8f36b84b146150373f1a699597b98f4f7
Reviewed-on: https://chromium-review.googlesource.com/1081895
Reviewed-by: agrieve <agrieve@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Commit-Queue: Samuel Huang <huangs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#563923}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4bcf50a6da995e67d62596a71533a85a5d63347c