Histogram Doc - Add Notes about Internal histograms.xml File
If this description is good, I'll make the analogous
change to the user actions doc.

Also added a link to histogram_macros.h at the top of the
"Coding" section in the histogram docs, and a link to the
user actions recording code at the top of "Coding" section
in the user actions doc.

Change-Id: I05b546fd1ff7e00046b9324c9d02ac5f85a7f0ab
Reviewed-on: https://chromium-review.googlesource.com/1079794
Commit-Queue: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#564618}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 159c389734bbb47ab87110f09385f41142cfe678