Notify the browser process about a bloated renderer via the GRC.
The renderer sends a RendererIsBloated event to the GRC service.
The GRC turns the event into a PageIsBloated signal and sends the
signal to the tab manager in the browser process. The tab manager
then reloads the bloated tab.

Subsequent CL will add a notification from the tab manager to the
web contents, so that it can show an inforbar after the reload.

Bug: 835806
Change-Id: I23d0987094f1c6864534abeb57f8cce3273172ff
Reviewed-on: https://chromium-review.googlesource.com/1073419
Commit-Queue: Ulan Degenbaev <ulan@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#565680}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e7d18c79fef2898cdc31d922b76097b27e078400