Added 'component' attribute to symbol and corresponding archive/group code
supersize now tracks data from COMPONENT files for symbols, along with lookups
for that info. The COMPONENT is resolved by checking the directory of the
current file for an OWNERS file. If it is not found there, then the program
moves to the parent directory, and so on, until the root of the file system is
reached.

Bug: 847576
Change-Id: If0d2ab8936614b4933b6850c082d7f116e40685d
Reviewed-on: https://chromium-review.googlesource.com/1080401
Commit-Queue: Tiger Oakes <tigero@google.com>
Reviewed-by: agrieve <agrieve@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#566472}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 050f837e45dc982d2531f51ed204eee7f34c0e2d