Declarative Net Request: Replace usages of 'blacklist' and 'whitelist'.
This CL replaces the usages of 'blacklist' and 'whitelist' in the Declarative
Net Request code. It introduces the following changes to the public API:
  - RuleActionType blacklist -> RuleActionType block
  - RuleActionType whitelist -> RuleActionType allow
  - MAX_NUMBER_OF_WHITELISTED_PAGES -> MAX_NUMBER_OF_ALLOWED_PAGES
  - addWhitelistedPages -> addAllowedPages
  - removeWhitelistedPages -> removeAllowedPages
  - getWhitelistedPages -> getAllowedPages

As a result of this change, existing extensions using the Declarative Net
Request API will need to be updated. Usages of these terms in enums and
preference keys are kept as-is.

BUG=850774, 842296

Change-Id: Id4aaac5ec3123fd56a1591e22025a455d7cb6941
Reviewed-on: https://chromium-review.googlesource.com/1094141
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#567858}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2860433068296b76190c08acdf857a0a5d4cabc3