Support launching Crostini apps from the File Manager
This CL adds support for launching Crostini apps from the File Manager.
For the moment, we only support files which are already in the Crostini
container, but in the future we will support external files as well.

Support for display app icons will be added in a separate patch, but as
it will require FindCrostiniTasks to be asynchronous, it is made to take
a callback to run upon completion.

Bug: 822513
Change-Id: I257b482991ab8cf2e4d0ef5be18d0e33205df4e3
Reviewed-on: https://chromium-review.googlesource.com/1100606
Commit-Queue: Timothy Loh <timloh@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#569171}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d11a5c2a75d6c0e85227c3c37dd42ccda553e236