Make IsDisturbed take an ExceptionState parameter
Modify blink::ReadableStreamOperations::IsDisturbed() to take an
ExceptionState& parameter. Modify all callers so that the ExceptionState
is passed through all the way from the bindings layer. Where
IsDisturbed() is called from code that runs asynchronously, the
context_type, interface_name and property_name are cached and used to
create a new ExceptionState object in the asynchronous context.

Also change methods in the implementation of the payment API to use the
cached context_type, interface_name and property_name instead of
hard-coding the values.

Add the extended attribute "RaisesException" to all the asynchronous
body-reading methods in body.idl so that they receive an ExceptionState&
parameter which is used to handle exceptions thrown when calling
IsDisturbed() inside RejectInvalidConsumption().
RejectInvalidConsumption() now always rejects by throwing an exception
and no longer returns a promise.

An exception-ignoring IsBodyUsedForDCheck() method is added which
doesn't change the exception state.

Also add the [RaisesException] attribute to the bodyUsed accessor. Since
it now throws if it has ever failed, fix the
call-extra-crash-is-disturbed.html layout test to reflect that.

Bug: 853189
Change-Id: Ib6df486a5f4ea7c58ecc123a716ee76ace0c92ed
Reviewed-on: https://chromium-review.googlesource.com/1107578
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Jinho Bang <jinho.bang@samsung.com>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#570621}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0598444f7467071501c71290248bf8758d501fde