Reland of: Add mojom call to send resource loading hints from browser to renderer.
The mojom call is send only when resource loading hints feature is
enabled and the connection is slow.

This is a reland of the CL that got reverted due to flaky
browsertest. This reland is quite similar to the reverted CL,
except for some changes in the browsertest.

PS#1 is the original CL that got reverted.

The fix is to retry fetching of histograms from the renderers.
Note that previews_browsertest are currently enabled only on Android.
So, it's possible that the same flakiness bug is affecting the
browser tests on Linux in this CL.

Bug: 854666
Cq-Include-Trybots: luci.chromium.try:linux_mojo
Change-Id: I753e488bef7318e43530a5594a94b24bd09295bb
TBR: rkaplow@chromium.org, dcheng@chromium.org, pkasting@chromium.org, kinuko@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/1131918
Commit-Queue: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#574419}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc97029f96fedad41b8600011d6c155810e50fe6