Add histogram to measure usage of '#' in WebView loadData APIs
The goal is to compare how often this histogram is logged with the
existing kDataUriHasOctothorpe UseCounter for WebView, in order to
determine what API path we might want to modify when doing a compat
fix. The belief is that the histogram and UseCounter should be
roughly comparable because the histogram will only be recorded at most
once per site load (because the loadData APIs cause a site load).

Bug: 823666
Change-Id: I0ce63b8c86bc3bcd02d3b10ab7a38112fb520439
Reviewed-on: https://chromium-review.googlesource.com/1112045
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Reviewed-by: Robert Kaplow (slow) <rkaplow@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#575268}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 73716e905be82c0f58474a25c213382f3993b43d