Reland "Forms: Add a UseCounter for opening popups while a file chooser is opening"
We'd like to fix a confusing UI issue, and to know the impact
of a possible fix.

This relands #575575, which caused node leaks.
If a document is navigated out while a file chooser is opening,
ChromeClientImpl holded a strong reference to a FileInputType and the
whole document leaked.
This CL changes the reference from strong one to weak one.

Bug: 637639
Change-Id: I5dccade80e433ac53de4bb205abab58380b0a5f2
Reviewed-on: https://chromium-review.googlesource.com/1140339
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#575923}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6bfe210aa3c20ebf1db6ce788991f24e1eace4c7