Mark automated downloads-related user action as not user triggered.
The action happens whenever the DownloadActivity is resumed, which
can happen multiple times when it is triggered. This leads to the
dashboard being cluttered with many of these events, that are
not really helpful notions of user actions.

BUG=786102

Change-Id: I534d2fa8f147281b619f8f10085a0aab8e3815bf
Reviewed-on: https://chromium-review.googlesource.com/1138723
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Theresa <twellington@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#576122}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 11e89159c458fb777343fd59e8da24306524a169